Simplify constexpr steps option configuration
Set compiler option to allow higher number of steps to be performed
globally for the whole project. This is easier mechanism that setting it
separately for each target. The value is also set to be large enough for
all known needs. Previously all targets had their own minimal but high
enough value to pass compilation. However, the value itself doesn't
affect the generated binaries and it is therefore not beneficial to
attempt minimizing the value as was done previously.