Add initial dev container with GCC
Add configuration for dev container with GCC 12. This configuration
contains also conan and triggers conan to install dependencies with
specific profile when attaching to the container.

CMakePresets is also updated to contain initial GCC support.

However, GCC 12 is not able to compile the source code of this project
as it doesn't yet support all used C++23 features. Therefore, this dev
container isn't yet very useful.