Update devcontainer with clang 17
Update devcontainer to include clang 17 and to install conan
dependencies for it during container creation.

This gets the container closer to being able to build the repository.
However, the build fails still due to libc++ not supporting necessary
C++23 features. `std::views::slide` being the first source of error
currently.