Convert devcontainer to Alpine for newer GCC
Alpine provides easy way to install GCC 13. Otherwise it would need
to be compiled from sources as there are no precompiled packages for
Debian nor Ubuntu.
Now the devcontainer can be used with both GCC 13 and LLVM 17.

However, both GCC 13 and LLVM 17 are not able to build the project
yet due to missing C++23 features.