devcontainer: Fix updateContent and conan variants
Previously the call to task looked as if it would install dependencies
with both Debug and Release variants for GCC and Clang. However, the
call to task was incorrect and only the last value of BUILD_TYPE was
actually used.

This commit fixes the call to task so that it will utilize taskfile
itself to specify necessary arguments. This also makes the dependency
install to run in parallel for all variants speeding up the process.