Disable invalid code using views::zip
Initial code contained a buggy implementation using views::zip and
it was only enabled if compiler supported that. However, at that time,
even MSVC didn't support it and it wasn't possible to ensure that it
even compiled or worked. Nowadays, MSVC does support it, but the
implementation doesn't compile.

This commit disables the buggy implementation even when compiler
supports it and adds a note about it needing to be fixed.