devcontainer: Clang with libstdc++11
Change the Clang Conan profile in devcontainer to use libstdc++11
instead of libc++ as the standard library. Libstdc++11 includes support
for all C++23 features used by this project unlike libc++.

This change allows clang to be used as the compiler in the devcontainer
and fixes the build errors that were occurring previously when using
clang.